Fundamental overhaul of the engine architecture to remove legacy limits.
- Transitioned the VM and numeric core from 32-bit float to 64-bit double.
- Expanded local, argument, and global indices to 32-bit (Int32), removing the 255-variable limit.
- Upgraded GameObject IDs to 64-bit (long) for massive scale support.
- Implemented full 64-bit bitwise operation support.
- Added DynamicLimitService to scale VM stack and global limits based on system RAM.
- Updated BytecodeOptimizer and networking serialization for the new formats.
- Refactored 170+ unit tests to ensure architectural stability.

+++ Engine/Core/Objects/ObjectApi.cs
@@ -36,7 +36,7 @@ namespace Core.Api
             }
         }
 
-        public GameObject? GetObject(int id)
+        public GameObject? GetObject(long id)
         {
             using (_gameState.ReadLock())
             {
@@ -45,7 +45,7 @@ namespace Core.Api
             }
         }
 
-        public void DestroyObject(int id)
+        public void DestroyObject(long id)
         {
             using (_gameState.WriteLock())
             {
@@ -58,7 +58,7 @@ namespace Core.Api
             }
         }
 
-        public void MoveObject(int id, int x, int y, int z)
+        public void MoveObject(long id, int x, int y, int z)
         {
             using (_gameState.WriteLock())
             {
